+++ test/libsolidity/SolidityABIJSON.cpp
@@ -76,7 +76,7 @@ BOOST_AUTO_TEST_CASE(basic_test)
 		"name": "f",
 		"constant": false,
 		"payable" : false,
-		"statemutability": "nonpayable",
+		"stateMutability": "nonpayable",
 		"type": "function",
 		"inputs": [
 		{
@@ -120,7 +120,7 @@ BOOST_AUTO_TEST_CASE(multiple_methods)
 		"name": "f",
 		"constant": false,
 		"payable" : false,
-		"statemutability": "nonpayable",
+		"stateMutability": "nonpayable",
 		"type": "function",
 		"inputs": [
 		{
@@ -139,7 +139,7 @@ BOOST_AUTO_TEST_CASE(multiple_methods)
 		"name": "g",
 		"constant": false,
 		"payable" : false,
-		"statemutability": "nonpayable",
+		"stateMutability": "nonpayable",
 		"type": "function",
 		"inputs": [
 		{
@@ -172,7 +172,7 @@ BOOST_AUTO_TEST_CASE(multiple_params)
 		"name": "f",
 		"constant": false,
 		"payable" : false,
-		"statemutability": "nonpayable",
+		"stateMutability": "nonpayable",
 		"type": "function",
 		"inputs": [
 		{
@@ -211,7 +211,7 @@ BOOST_AUTO_TEST_CASE(multiple_methods_order)
 		"name": "c",
 		"constant": false,
 		"payable" : false,
-		"statemutability": "nonpayable",
+		"stateMutability": "nonpayable",
 		"type": "function",
 		"inputs": [
 		{
@@ -230,7 +230,7 @@ BOOST_AUTO_TEST_CASE(multiple_methods_order)
 		"name": "f",
 		"constant": false,
 		"payable" : false,
-		"statemutability": "nonpayable",
+		"stateMutability": "nonpayable",
 		"type": "function",
 		"inputs": [
 		{
@@ -264,7 +264,7 @@ BOOST_AUTO_TEST_CASE(view_function)
 		"name": "foo",
 		"constant": false,
 		"payable" : false,
-		"statemutability": "nonpayable",
+		"stateMutability": "nonpayable",
 		"type": "function",
 		"inputs": [
 		{
@@ -287,7 +287,7 @@ BOOST_AUTO_TEST_CASE(view_function)
 		"name": "boo",
 		"constant": true,
 		"payable" : false,
-		"statemutability": "view",
+		"stateMutability": "view",
 		"type": "function",
 		"inputs": [{
 			"name": "a",
@@ -320,7 +320,7 @@ BOOST_AUTO_TEST_CASE(constant_function)
 		"name": "foo",
 		"constant": false,
 		"payable" : false,
-		"statemutability": "nonpayable",
+		"stateMutability": "nonpayable",
 		"type": "function",
 		"inputs": [
 		{
@@ -343,7 +343,7 @@ BOOST_AUTO_TEST_CASE(constant_function)
 		"name": "boo",
 		"constant": true,
 		"payable" : false,
-		"statemutability": "view",
+		"stateMutability": "view",
 		"type": "function",
 		"inputs": [{
 			"name": "a",
@@ -375,7 +375,7 @@ BOOST_AUTO_TEST_CASE(pure_function)
 		"name": "foo",
 		"constant": false,
 		"payable" : false,
-		"statemutability": "nonpayable",
+		"stateMutability": "nonpayable",
 		"type": "function",
 		"inputs": [
 		{
@@ -398,7 +398,7 @@ BOOST_AUTO_TEST_CASE(pure_function)
 		"name": "boo",
 		"constant": true,
 		"payable" : false,
-		"statemutability": "pure",
+		"stateMutability": "pure",
 		"type": "function",
 		"inputs": [{
 			"name": "a",
@@ -430,7 +430,7 @@ BOOST_AUTO_TEST_CASE(events)
 		"name": "f",
 		"constant": false,
 		"payable" : false,
-		"statemutability": "nonpayable",
+		"stateMutability": "nonpayable",
 		"type": "function",
 		"inputs": [
 		{
@@ -512,7 +512,7 @@ BOOST_AUTO_TEST_CASE(inherited)
 		"name": "baseFunction",
 		"constant": false,
 		"payable" : false,
-		"statemutability": "nonpayable",
+		"stateMutability": "nonpayable",
 		"type": "function",
 		"inputs":
 		[{
@@ -529,7 +529,7 @@ BOOST_AUTO_TEST_CASE(inherited)
 		"name": "derivedFunction",
 		"constant": false,
 		"payable" : false,
-		"statemutability": "nonpayable",
+		"stateMutability": "nonpayable",
 		"type": "function",
 		"inputs":
 		[{
@@ -585,7 +585,7 @@ BOOST_AUTO_TEST_CASE(empty_name_input_parameter_with_named_one)
 		"name": "f",
 		"constant": false,
 		"payable" : false,
-		"statemutability": "nonpayable",
+		"stateMutability": "nonpayable",
 		"type": "function",
 		"inputs": [
 		{
@@ -628,7 +628,7 @@ BOOST_AUTO_TEST_CASE(empty_name_return_parameter)
 		"name": "f",
 		"constant": false,
 		"payable" : false,
-		"statemutability": "nonpayable",
+		"stateMutability": "nonpayable",
 		"type": "function",
 		"inputs": [
 		{
@@ -672,7 +672,7 @@ BOOST_AUTO_TEST_CASE(constructor_abi)
 			}
 		],
 		"payable": false,
-		"statemutability": "nonpayable",
+		"stateMutability": "nonpayable",
 		"type": "constructor"
 	}
 	])";
@@ -704,7 +704,7 @@ BOOST_AUTO_TEST_CASE(payable_constructor_abi)
 			}
 		],
 		"payable": true,
-		"statemutability": "payable",
+		"stateMutability": "payable",
 		"type": "constructor"
 	}
 	])";
@@ -730,7 +730,7 @@ BOOST_AUTO_TEST_CASE(return_param_in_abi)
 		{
 			"constant" : false,
 			"payable" : false,
-			"statemutability": "nonpayable",
+			"stateMutability": "nonpayable",
 			"inputs" : [],
 			"name" : "ret",
 			"outputs" : [
@@ -749,7 +749,7 @@ BOOST_AUTO_TEST_CASE(return_param_in_abi)
 				}
 			],
 			"payable": false,
-			"statemutability": "nonpayable",
+			"stateMutability": "nonpayable",
 			"type": "constructor"
 		}
 	]
@@ -771,7 +771,7 @@ BOOST_AUTO_TEST_CASE(strings_and_arrays)
 		{
 			"constant" : false,
 			"payable" : false,
-			"statemutability": "nonpayable",
+			"stateMutability": "nonpayable",
 			"name": "f",
 			"inputs": [
 				{ "name": "a", "type": "string" },
@@ -800,7 +800,7 @@ BOOST_AUTO_TEST_CASE(library_function)
 		{
 			"constant" : false,
 			"payable" : false,
-			"statemutability": "nonpayable",
+			"stateMutability": "nonpayable",
 			"name": "f",
 			"inputs": [
 				{ "name": "b", "type": "test.StructType storage" },
@@ -830,7 +830,7 @@ BOOST_AUTO_TEST_CASE(include_fallback_function)
 	[
 		{
 			"payable": false,
-			"statemutability": "nonpayable",
+			"stateMutability": "nonpayable",
 			"type" : "fallback"
 		}
 	]
@@ -852,7 +852,7 @@ BOOST_AUTO_TEST_CASE(payable_function)
 		{
 			"constant" : false,
 			"payable": false,
-			"statemutability": "nonpayable",
+			"stateMutability": "nonpayable",
 			"inputs": [],
 			"name": "f",
 			"outputs": [],
@@ -861,7 +861,7 @@ BOOST_AUTO_TEST_CASE(payable_function)
 		{
 			"constant" : false,
 			"payable": true,
-			"statemutability": "payable",
+			"stateMutability": "payable",
 			"inputs": [],
 			"name": "g",
 			"outputs": [],
@@ -884,7 +884,7 @@ BOOST_AUTO_TEST_CASE(payable_fallback_function)
 	[
 		{
 			"payable": true,
-			"statemutability": "payable",
+			"stateMutability": "payable",
 			"type" : "fallback"
 		}
 	]
@@ -905,7 +905,7 @@ BOOST_AUTO_TEST_CASE(function_type)
 	{
 		"constant" : false,
 		"payable": false,
-		"statemutability": "nonpayable",
+		"stateMutability": "nonpayable",
 		"inputs": [{
 			"name": "x",
 			"type": "function"